move useAddysSummary to createQueryStore
