+++ src/hooks/useWalletTransactionCounts.ts
@@ -2,7 +2,7 @@ import { AllRainbowWallets } from '@/model/wallet';
 import { useMemo } from 'react';
 import { Address } from 'viem';
 import useAccountSettings from './useAccountSettings';
-import { useAddysSummary } from '@/resources/summary/summary';
+import { useAddysSummary } from '@/resources/addys/summary';
 
 const QUERY_CONFIG = {
   staleTime: 60_000, // 1 minute
