+++ src/resources/addys/client.ts
@@ -3,12 +3,13 @@ import { ADDYS_API_KEY, ADDYS_BASE_URL } from 'react-native-dotenv';
 
 let addysHttp: RainbowFetchClient | undefined;
 
-export const getAddysHttpClient = () => {
+export const getAddysHttpClient = ({ abortController }: { abortController?: AbortController | null } = {}) => {
   const clientUrl = addysHttp?.baseURL;
   const baseUrl = ADDYS_BASE_URL;
   if (!addysHttp || clientUrl !== baseUrl) {
     addysHttp = new RainbowFetchClient({
       baseURL: ADDYS_BASE_URL,
+      abortController,
       headers: {
         Authorization: `Bearer ${ADDYS_API_KEY}`,
       },
